pyproject.toml
[tool.poetry]
name = "observml"
version = "1.0.0"
description = "Modular machine learning for anomaly detection"
authors = ["ObservML Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = ">=1.26.0,<2.2.0"
pandas = "^2.2.3"
pyyaml = "^6.0.2"
pika = "^1.3.2"
hydra-core = "^1.3.2"
openpyxl = "^3.1.5"
mlflow = "^2.22.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
plotly = "^6.0.1"
scikit-learn = "^1.6.1"
tensorflow = "^2.19.0"
prophet = "^1.1.6"
hmmlearn = "^0.3.3"
pyts = "^0.13.0"
pmdarima = "^2.0.4"
distinctipy = "^1.2.3"
pgmpy = "^0.1.25"
mchmm = "^0.4.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
pytest-mock = "^3.14.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
